{edx_dl, parsing}: Move function to separate module.
With this, we move a pure function that is meant to parse things to a module
whose purpose is to simply parse text.
