+++ edx_dl/edx_dl.py
@@ -8,7 +8,6 @@ import re
 import sys
 
 from collections import namedtuple
-from datetime import timedelta, datetime
 from functools import partial
 from multiprocessing.dummy import Pool as ThreadPool
 
@@ -26,6 +25,7 @@ from six.moves.urllib.request import (
 )
 
 from .compat import compat_print
+from .parsing import edx_json2srt
 from .utils import (
     directory_name,
     execute_command,
@@ -211,31 +211,6 @@ def get_available_sections(url, headers):
                         subsections=_make_subsections(section_soup))
                 for i, section_soup in enumerate(sections_soup, 1)]
     return sections
-
-
-def edx_json2srt(o):
-    """
-    Transform the dict 'o' into the srt subtitles format
-    """
-    BASE_TIME = datetime(1, 1, 1)
-    output = []
-
-    for i, (s, e, t) in enumerate(zip(o['start'], o['end'], o['text'])):
-        if t == '':
-            continue
-
-        output.append(str(i) + '\n')
-
-        s = BASE_TIME + timedelta(seconds=s/1000.)
-        e = BASE_TIME + timedelta(seconds=e/1000.)
-        time_range = "%02d:%02d:%02d,%03d --> %02d:%02d:%02d,%03d\n" % \
-                     (s.hour, s.minute, s.second, s.microsecond/1000,
-                      e.hour, e.minute, e.second, e.microsecond/1000)
-
-        output.append(time_range)
-        output.append(t + "\n\n")
-
-    return ''.join(output)
 
 
 def edx_get_subtitle(url, headers):
