+++ edx_dl/parsing.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+
+from datetime import timedelta, datetime
+
+
+def edx_json2srt(o):
+    """
+    Transform the dict 'o' into the srt subtitles format
+    """
+    BASE_TIME = datetime(1, 1, 1)
+    output = []
+
+    for i, (s, e, t) in enumerate(zip(o['start'], o['end'], o['text'])):
+        if t == '':
+            continue
+
+        output.append(str(i) + '\n')
+
+        s = BASE_TIME + timedelta(seconds=s/1000.)
+        e = BASE_TIME + timedelta(seconds=e/1000.)
+        time_range = "%02d:%02d:%02d,%03d --> %02d:%02d:%02d,%03d\n" % \
+                     (s.hour, s.minute, s.second, s.microsecond/1000,
+                      e.hour, e.minute, e.second, e.microsecond/1000)
+
+        output.append(time_range)
+        output.append(t + "\n\n")
+
+    return ''.join(output)
